Add timetable to schedule
I used a js library [0] to create the timetable.

On one hand the data is taken from the data.js file so in the future if
we want to get the data from the database it shouldn't be too diffucult.

On the other hand an HTML table would be simpler to understand.

[0] https://github.com/codrops/Calendario