Common PHP and js file for the database query
I've added the /js/getDBTable.js file to the head.html file so it's
automatically included in every file.

To get data from the DB you have to:

1) Add a js file to the index page (like we do now with the
getCourse.js).

2) In that js file call getDBTable with first parameter equal to the name
of the table you want to get and the second parameter is the function that
will be run when the query is done. The first parameter of that function is
a JSON object on which you can iterate to add the info to the page.

The function as parameter is needed since we are getting the info from
the database asynchronously, this function is called callback;

Just to be clear this is the order in which the functions get called:

$(document).ready -> getDBTable -> $.ajax -> getDB.php -> getDBTable
callback function ;

The getDB.php script gets the table name from the POST data, for now it
only gets whole tables. When we need to add other mysql querys we can
just change the getDB.php file and everything else should countinue
working as long as we keep returning valid JSON objects;