getCourse move 2nd getDBTable to callback
getDBTable is asynchronous, so when you call it the
callback you pass to it will not be necessarily run right after.

When you put 2 getDBTable in the same function one after another you
can't be sure that the callback you passed to the first getDBTable will
be run before the callback you passed to the second getDBTable.

In the getCourse file let's call:
getDBTable("category" -> A
function(category     -> B (the first getDBTable callback)
getDBTable("course"   -> C
function(course       -> D (the second getDBTable callback)

They can run in any of these orders:

1) A -> B -> C -> D
2) A -> C -> B -> D
3) A -> C -> D -> B

In case 3 the categoryTable array won't exist when the function(course
callback is run and categoryTable[j] will still be undefined.

Thats why I moved getDBTable("course" to the function(category,
callback, now the run order is necessarily:

A -> B -> C -> D